refactor: добавить бизнес-логику в юзердао и валидацию в юзер-сервис

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -5,7 +5,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import ru.practicum.shareit.exception.EntityNotFoundException;
+import ru.practicum.shareit.exception.ValidationException;
 import ru.practicum.shareit.user.dao.UserDao;
+import ru.practicum.shareit.user.dao.UserDaoImpl;
+import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.model.User;
 
 import java.util.List;
@@ -21,26 +25,53 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User createUser(User user) {
-        return userDao.createUser(user);
+    public UserDto createUser(UserDto userDto) {
+        if (userDto.getName() == null || userDto.getName().isEmpty()) {
+            throw new ValidationException("Имя не должно быть пустым.");
+        } else if (userDto.getEmail() == null || userDto.getEmail().isEmpty()) {
+            throw new ValidationException("Адрес электронной почты не должен быть пустым.");
+        } else if (!userDto.getEmail().contains("@")) {
+            throw new ValidationException("Некорректный формат e-mail. Адрес электронной почты должен содержать символ '@'.");
+        }
+        return userDao.createUser(userDto);
     }
 
     @Override
-    public List<User> getAllUsers() {
+    public List<UserDto> getAllUsers() {
         return userDao.getAllUsers();
     }
 
     @Override
-    public User getUserById(long id) {
+    public UserDto getUserById(long id) {
+        if (!userDao.isContain(id)) {
+            throw new EntityNotFoundException("Пользователь с указанным id = " + id + " не найден.");
+        }
         return userDao.getUserById(id);
     }
     @Override
-    public User updateUser(long id) {
-       return userDao.updateUser(id);
+    public UserDto updateUser(long id, UserDto userDto) {
+        if (!userDao.isContain(id)) {
+            throw new EntityNotFoundException("Пользователь с указанным id = \" + id + \" не найден.");
+        }
+        if (userDto.getEmail() != null) {
+            if (!userDto.getEmail().contains("@")) {
+                throw new ValidationException("Некорректный формат e-mail. Адрес электронной почты должен содержать символ '@'.");
+            }
+        }
+        if (userDto.getName() != null) {
+            if (userDto.getName().equals("")) {
+                throw new ValidationException("Поле не должно быть пустым.");
+            }
+        }
+       return userDao.updateUser(id, userDto);
     }
 
     @Override
     public void deleteUser(long  id) {
+        if (!userDao.isContain(id)) {
+            throw new EntityNotFoundException("Пользователь с указанным id = " + id + " не найден.");
+        }
        userDao.deleteUser(id);
     }
+
 }
